# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing buildings and passes.",
        "title": "Service-propusk API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "172.17.17.145:8000",
    "basePath": "/",
    "paths": {
...
        "/users": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию о текущем пользователе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Обновить данные пользователя",
                "parameters": [
                    {
                        "description": "Обновленные данные пользователя",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userHandler.EditUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/userHandler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/userHandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/userHandler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает нового пользователя в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Зарегистрировать нового пользователя",
                "parameters": [
                    {
                        "description": "Данные нового пользователя",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userHandler.RegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/userHandler.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/userHandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/userHandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Аутентифицирует пользователя и возвращает токен доступа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Войти в систему",
                "parameters": [
                    {
                        "description": "Учетные данные пользователя",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userHandler.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/userHandler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/userHandler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/userHandler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "userHandler.Credentials": {
            "description": "Учетные данные для входа в систему",
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "description": "Email пользователя",
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "description": "Пароль пользователя",
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "userHandler.EditUserRequest": {
            "description": "Запрос на обновление данных пользователя",
            "type": "object",
            "properties": {
                "login": {
                    "description": "Новый email пользователя",
                    "type": "string",
                    "example": "newuser@example.com"
                },
                "password": {
                    "description": "Новый пароль пользователя",
                    "type": "string",
                    "example": "newpassword123"
                }
            }
        },
        "userHandler.ErrorResponse": {
            "description": "Структура ответа при возникновении ошибки",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Текст ошибки",
                    "type": "string",
                    "example": "Описание ошибки"
                }
            }
        },
        "userHandler.RegistrationRequest": {
            "description": "Запрос на регистрацию нового пользователя",
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "description": "Email пользователя",
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "description": "Пароль пользователя (минимум 6 символов)",
                    "type": "string",
                    "minLength": 6,
                    "example": "password123"
                }
            }
        }
    },
    "securityDefinitions": {
        "cookieAuth": {
            "description": "Enter your JWT token value",
            "type": "apiKey",
            "name": "X-Access-Token",
            "in": "header"
        }
    }
}